btrfs-progs: move free-space-tree.c to kernel-shared/
