+++ kernel-shared/free-space-cache.h
