+++ kernel-shared/free-space-tree.c
@@ -18,7 +18,7 @@
 
 #include "ctree.h"
 #include "disk-io.h"
-#include "free-space-cache.h"
+#include "kernel-shared/free-space-cache.h"
 #include "kernel-shared/free-space-tree.h"
 #include "volumes.h"
 #include "transaction.h"
